Trabalho-2-elevador/requirements.txt
pyserial    ==    3.5
smbus2      ==    0.4.3
RPI.BME280  ==    0.2.4
RPi.GPIO    ==    0.7.1
smbus       ==    1.1.post2
adafruit-ssd1306
